Start adopting the cli workload install command
Instead of manually finding and installing all of the SDK packs, we should instead start using `dotnet workload install` which now uses `--skip-manifest-update` to ensure all the packs are installed for the workload, if the dotnet sdk version is 6.0-preview.6 or newer.

This older code to install workloads can go away in the future.