Rework workload logic for missing packs
Requesting the suggested workloads for a given set of missing packs works on windows but was causing issues on mac, not sure why.

Instead let's just assume we should run dotnet workload install --skip-manifest-update for ALL the workloads in the maui-check manifest when one or more packs are missing since it's pretty cheap/safe operation to do anyway.